+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### Features
 1. [#172](https://github.com/influxdata/influxdb-client-java/pull/172): flux-dsl: added `to` function without `org` parameter
 
+### Bug Fixes
+1. [#173](https://github.com/influxdata/influxdb-client-java/pull/173): Query error could be after _success_ table
+
 ## 1.13.0 [2020-10-30]
 
 ### Features